+++ internal/app/displays/publication_details.go
@@ -1,9 +1,9 @@
 package displays
 
 import (
-	"github.com/ugent-library/biblio-backend/internal/locale"
-	"github.com/ugent-library/biblio-backend/internal/models"
-	"github.com/ugent-library/biblio-backend/internal/render/display"
+	"github.com/ugent-library/biblio-backoffice/internal/locale"
+	"github.com/ugent-library/biblio-backoffice/internal/models"
+	"github.com/ugent-library/biblio-backoffice/internal/render/display"
 )
 
 func PublicationDetails(user *models.User, l *locale.Locale, p *models.Publication) *display.Display {