+++ internal/backends/excel/dataset/dataset_list_exporter.go
@@ -5,10 +5,10 @@ import (
 	"io"
 	"strings"
 
-	"github.com/ugent-library/biblio-backend/internal/backends"
-	"github.com/ugent-library/biblio-backend/internal/backends/excel"
-	"github.com/ugent-library/biblio-backend/internal/models"
-	internal_time "github.com/ugent-library/biblio-backend/internal/time"
+	"github.com/ugent-library/biblio-backoffice/internal/backends"
+	"github.com/ugent-library/biblio-backoffice/internal/backends/excel"
+	"github.com/ugent-library/biblio-backoffice/internal/models"
+	internal_time "github.com/ugent-library/biblio-backoffice/internal/time"
 )
 
 const sep = " ; "
@@ -86,7 +86,11 @@ func (x *xlsx) datasetToRow(ds *models.Dataset) []string {
 		{
 			values := []string{}
 			for _, contributor := range contributors {
-				values = append(values, contributor.FullName)
+				fullName := contributor.FullName
+				if fullName == "" {
+					fullName = contributor.FirstName + " " + contributor.LastName
+				}
+				values = append(values, fullName)
 			}
 			m[role] = strings.Join(values, sep)
 		}
@@ -102,7 +106,11 @@ func (x *xlsx) datasetToRow(ds *models.Dataset) []string {
 				}
 				//full_name (<ugent_id>)
 				//full_name (<ugent_id>@<department.0.id>)
-				val := fmt.Sprintf("%s (%s%s)", contributor.FullName, contributor.UGentID[0], group)
+				fullName := contributor.FullName
+				if fullName == "" {
+					fullName = contributor.FirstName + " " + contributor.LastName
+				}
+				val := fmt.Sprintf("%s (%s%s)", fullName, contributor.UGentID[0], group)
 				values = append(values, val)
 			}
 			m["ugent_"+role] = strings.Join(values, sep)