+++ internal/commands/index.go
@@ -5,7 +5,7 @@ import (
 	"sync"
 
 	"github.com/spf13/cobra"
-	"github.com/ugent-library/biblio-backend/internal/models"
+	"github.com/ugent-library/biblio-backoffice/internal/models"
 )
 
 func init() {